requirements.txt
python-binance>=1.0.17
pandas>=1.3.0
schedule>=1.1.0
ta>=0.10.2
python-dotenv>=1.0.0
requests>=2.26.0
